+++ lib/nginx_stage/generators/nginx_process_generator.rb
@@ -34,7 +34,14 @@ module NginxStage
     # Run the per-user NGINX process (exit quietly on success)
     add_hook :exec_nginx do
       if !skip_nginx
-        o, s = Open3.capture2e([NginxStage.nginx_bin, "(#{user})"], *NginxStage.nginx_args(user: user, signal: signal))
+        o, s = Open3.capture2e(
+          NginxStage.nginx_env(user: user),
+          [
+            NginxStage.nginx_bin,
+            "(#{user})"
+          ],
+          *NginxStage.nginx_args(user: user, signal: signal)
+        )
         s.success? ? exit : abort(o)
       end
     end